+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.4]
+### Fix
+- _eq_is_eq_data as False in workflow class definitions
+
+### Changed
+- introducing _eq_is_data_eq instead of _generic eq
+- __origin__ used in serialize_typing instead of _name. For typing type checking
+
 ### Removed
 - Compiled parts of package, switching to full open-source